pyproject.toml
[project]
name = "made"
authors = [
    { name="Shreshth Malik", email="shreshth@diffractive.ai" }
]
version = "0.1.0"
description = "Benchmark environments for closed-loop materials discovery"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ase>=3.25.0",
    "ase-ga>=0.2.0",
    "average-minimum-distance>=1.6.0",
    "chemeleon-dng @ git+https://github.com/hspark1212/chemeleon-dng.git",
    "dspy>=3.0.3",
    "hydra-core>=1.3.1",
    "kaleido>=1.0.0",
    "mace-torch>=0.3.14",
    "modal>=1.1.1",
    "mp-api>=0.45.8",
    "orb-models>=0.5.4",
    "pydantic>=2.0.0",
    "pymatgen>=2025.6.14",
    "scipy>=1.16.1",
    "smact>=3.2.0",
    "wandb>=0.21.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "jupyter>=1.1.1",
    "seaborn>=0.13.2",
]

[tool.ruff]
line-length = 88
target-version = "py312"


[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "UP", "C4", "SIM", "PIE"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["made"]

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
